🐛 Made some eval necessary fixes      - Fixed issue with ft_atol returning an int and not a long      - If there is only one item in the list it should not do anything, not return an error.      - If the first item in the list is a + or a -, and there is no number after it, it needs to error.

+++ Codam Core/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: llourens <llourens@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/10/09 11:30:38 by llourens      #+#    #+#                 */
-/*   Updated: 2025/01/10 18:43:19 by llourens      ########   odam.nl         */
+/*   Updated: 2025/02/10 17:53:21 by llourens      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-int		ft_atol(const char *nptr);
+long	ft_atol(const char *nptr);
 int		ft_atoi(const char *nptr);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t num, size_t size);
